Factored out MultiClickGuard class
