+++ collect_app/src/main/java/org/odk/collect/android/fragments/Camera2VideoFragment.java
@@ -47,8 +47,8 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import org.odk.collect.android.R;
-import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.fragments.dialogs.ErrorDialog;
+import org.odk.collect.android.utilities.MultiClickGuard;
 import org.odk.collect.android.utilities.CameraUtils;
 import org.odk.collect.android.utilities.ToastUtils;
 
@@ -294,7 +294,7 @@ public class Camera2VideoFragment extends Fragment
     @Override
     public void onClick(View view) {
         if (view.getId() == R.id.texture) {
-            if (Collect.allowClick(getClass().getName())) { // avoid multiple quick taps that may cause various problems
+            if (MultiClickGuard.allowClick(getClass().getName())) { // avoid multiple quick taps that may cause various problems
                 if (isRecordingVideo) {
                     textureView.setClickable(false);
                     stopRecordingVideo();
